EVG-7524: task status dropdown updates url statuses param
