+++ src/hooks/index.ts
@@ -4,6 +4,7 @@ import { useDefaultPath } from "hooks/useDefaultPath";
 import { useDisableTableSortersIfLoading } from "hooks/useDisableTableSortersIfLoading";
 import { useOnClickOutside } from "hooks/useOnClickOutside";
 import { useFilterInputChangeHandler } from "hooks/useFilterInputChangeHandler";
+import { useStatusesFilter } from "hooks/useStatusesFilter";
 
 export const usePrevious = <T>(state: T): T | undefined => {
   const ref = useRef<T>();
@@ -18,5 +19,6 @@ export {
   useDefaultPath,
   useDisableTableSortersIfLoading,
   useOnClickOutside,
-  useFilterInputChangeHandler
+  useFilterInputChangeHandler,
+  useStatusesFilter
 };
