+++ src/pages/task/testsTable/StatusSelector.tsx
@@ -1,22 +1,11 @@
 import React from "react";
-import { useLocation, useHistory } from "react-router-dom";
-import queryString from "query-string";
-import { RequiredQueryParams } from "types/task";
 import { TestStatus } from "types/task";
 import { TreeSelect } from "components/TreeSelect";
+import { useStatusesFilter } from "hooks";
+import { RequiredQueryParams } from "types/task";
 
 export const StatusSelector = () => {
-  const { pathname, search } = useLocation();
-  const { replace } = useHistory();
-  const value = useQueryParamStatuses(search);
-
-  const onChange = (updatedValue: [string]) => {
-    const parsed = queryString.parse(search, { arrayFormat });
-    parsed[RequiredQueryParams.Statuses] = updatedValue;
-    const nextQueryParams = queryString.stringify(parsed, { arrayFormat });
-    replace(`${pathname}?${nextQueryParams}`);
-  };
-
+  const [value, onChange] = useStatusesFilter(RequiredQueryParams.Statuses);
   return (
     <TreeSelect
       onChange={onChange}
@@ -27,8 +16,6 @@ export const StatusSelector = () => {
     />
   );
 };
-
-const arrayFormat = "comma";
 
 const treeData = [
   {
@@ -57,9 +44,3 @@ const treeData = [
     key: TestStatus.SilentFail
   }
 ];
-
-const useQueryParamStatuses = (search: string) => {
-  const parsed = queryString.parse(search, { arrayFormat });
-  const statuses = parsed[RequiredQueryParams.Statuses];
-  return Array.isArray(statuses) ? statuses : [statuses].filter(v => v);
-};
